separada la funcion de expansion de variable en una nueva funcion, hacer la funcion is_quote no static y cambiado el Makefile para eso, junto con comentar en el main la funcion print_cmd_list(temp)
